Minor documentation change
Hey folks! 

I was starting to use this  and noticed the steps to download the binary in the readme don't work - for whatever reason it looks like the 0.5.3 release didn't create binaries. I bumped the version to 0.5.4, and also tweaked the chmod command. (Looks like it might have had a typo - I just confirmed both commands work on a linux vm.)

Cheers!